Update and rename PHPUnit configuration file
In this commit, the `phpunit.xml` file is removed and replaced by a new `phpunit.xml.dist` file. This change aligns with the PHPUnit's best practices of using distributed configuration file and makes the testing process more adaptable across different environments. Several PHPUnit configurations are enhanced, such as `backupGlobals`, `stopOnFailure`, `executionOrder`, to improve the testing efficiency and reliability. The report and logging output directory is also updated to "build" directory for better organization of test results.