Update workflow triggers for push and PR events
Updated the branch triggers in the GitHub workflow files 'pest-types.yml' and 'pest.yml' to include 'master' branch in addition to 'main'. This ensures the workflows are triggered on both main and master branches during push and pull request events. This change offers flexibility to users using either 'main' or 'master' as their default branch.