Update tokenizer to user filter type enums instead of pure strings
Modified the tokenizer logic such that it now accepts an input argument of TokenizerFilterType instead of a string for specifying word filters and separation patterns. This makes the code more robust and readable, as we are now using explicit enum values instead of raw strings.

Additionally, all the corresponding usage in unit tests and README have also been adjusted to reflect this change. This would make it easier to add new filter types in future without modifying all the associated code.

The same change has been extended to the TokenizerFilter component as well, and a redundant long list of string constants in it has been removed since they are now represented by the enum.