Change private attributes and methods to protected
The access modifiers of several attributes and methods in `Tokenizer.php` have been changed from private to protected. This change enables subclasses to access these members and improves the extensibility of code. The attributes and methods include `$wordFilters`, `$wordSeparationPatterns`, `$toLowercase`, `getFilterPatterns()`, `getFilterReplacements()`, and `toLowercaseTokens()`. Also, a return type specification was added to the `addWordSeparatorPattern` method.